Windows libnetwork integration
