+++ daemon/daemon_windows.go
@@ -9,6 +9,7 @@ import (
 	"runtime"
 	"strings"
 
+	"github.com/Microsoft/hcsshim"
 	"github.com/Sirupsen/logrus"
 	"github.com/docker/docker/container"
 	"github.com/docker/docker/daemon/graphdriver"
@@ -16,6 +17,7 @@ import (
 	"github.com/docker/docker/image"
 	"github.com/docker/docker/layer"
 	"github.com/docker/docker/reference"
+	"github.com/docker/docker/runconfig"
 	containertypes "github.com/docker/engine-api/types/container"
 	// register the windows graph driver
 	"github.com/docker/docker/daemon/graphdriver/windows"
@@ -23,11 +25,15 @@ import (
 	"github.com/docker/docker/pkg/system"
 	"github.com/docker/libnetwork"
 	nwconfig "github.com/docker/libnetwork/config"
+	winlibnetwork "github.com/docker/libnetwork/drivers/windows"
+	"github.com/docker/libnetwork/netlabel"
+	"github.com/docker/libnetwork/options"
 	blkiodev "github.com/opencontainers/runc/libcontainer/configs"
 )
 
 const (
 	defaultVirtualSwitch = "Virtual Switch"
+	defaultNetworkSpace  = "172.16.0.0/12"
 	platformSupported    = true
 	windowsMinCPUShares  = 1
 	windowsMaxCPUShares  = 10000
@@ -125,16 +131,154 @@ func configureMaxThreads(config *Config) error {
 	return nil
 }
 
-func isBridgeNetworkDisabled(config *Config) bool {
-	return false
+func (daemon *Daemon) initNetworkController(config *Config) (libnetwork.NetworkController, error) {
+	// TODO Windows: Remove this check once TP4 is no longer supported
+	osv, err := system.GetOSVersion()
+	if err != nil {
+		return nil, err
+	}
+
+	if osv.Build < 14260 {
+		// Set the name of the virtual switch if not specified by -b on daemon start
+		if config.bridgeConfig.Iface == "" {
+			config.bridgeConfig.Iface = defaultVirtualSwitch
+		}
+		logrus.Warnf("Network controller is not supported by the current platform build version")
+		return nil, nil
+	}
+
+	netOptions, err := daemon.networkOptions(config)
+	if err != nil {
+		return nil, err
+	}
+	controller, err := libnetwork.New(netOptions...)
+	if err != nil {
+		return nil, fmt.Errorf("error obtaining controller instance: %v", err)
+	}
+
+	hnsresponse, err := hcsshim.HNSListNetworkRequest("GET", "", "")
+	if err != nil {
+		return nil, err
+	}
+
+	// Remove networks not present in HNS
+	for _, v := range controller.Networks() {
+		options := v.Info().DriverOptions()
+		hnsid := options[winlibnetwork.HNSID]
+		found := false
+
+		for _, v := range hnsresponse {
+			if v.Id == hnsid {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			err = v.Delete()
+			if err != nil {
+				return nil, err
+			}
+		}
+	}
+
+	_, err = controller.NewNetwork("null", "none", libnetwork.NetworkOptionPersist(false))
+	if err != nil {
+		return nil, err
+	}
+
+	// discover and add HNS networks to windows
+	// network that exist are removed and added again
+	for _, v := range hnsresponse {
+		var n libnetwork.Network
+		s := func(current libnetwork.Network) bool {
+			options := current.Info().DriverOptions()
+			if options[winlibnetwork.HNSID] == v.Id {
+				n = current
+				return true
+			}
+			return false
+		}
+
+		controller.WalkNetworks(s)
+		if n != nil {
+			v.Name = n.Name()
+			n.Delete()
+		}
+
+		netOption := map[string]string{
+			winlibnetwork.NetworkName: v.Name,
+			winlibnetwork.HNSID:       v.Id,
+		}
+
+		v4Conf := []*libnetwork.IpamConf{}
+		for _, subnet := range v.Subnets {
+			ipamV4Conf := libnetwork.IpamConf{}
+			ipamV4Conf.PreferredPool = subnet.AddressPrefix
+			ipamV4Conf.Gateway = subnet.GatewayAddress
+			v4Conf = append(v4Conf, &ipamV4Conf)
+		}
+
+		name := v.Name
+		// There is only one nat network supported in windows.
+		// If it exists with a different name add it as the default name
+		if runconfig.DefaultDaemonNetworkMode() == containertypes.NetworkMode(strings.ToLower(v.Type)) {
+			name = runconfig.DefaultDaemonNetworkMode().NetworkName()
+		}
+
+		v6Conf := []*libnetwork.IpamConf{}
+		_, err := controller.NewNetwork(strings.ToLower(v.Type), name,
+			libnetwork.NetworkOptionGeneric(options.Generic{
+				netlabel.GenericData: netOption,
+			}),
+			libnetwork.NetworkOptionIpam("default", "", v4Conf, v6Conf, nil),
+		)
+
+		if err != nil {
+			logrus.Errorf("Error occurred when creating network %v", err)
+		}
+	}
+
+	if !config.DisableBridge {
+		// Initialize default driver "bridge"
+		if err := initBridgeDriver(controller, config); err != nil {
+			return nil, err
+		}
+	}
+
+	return controller, nil
 }
 
-func (daemon *Daemon) initNetworkController(config *Config) (libnetwork.NetworkController, error) {
-	// Set the name of the virtual switch if not specified by -b on daemon start
-	if config.bridgeConfig.VirtualSwitchName == "" {
-		config.bridgeConfig.VirtualSwitchName = defaultVirtualSwitch
+func initBridgeDriver(controller libnetwork.NetworkController, config *Config) error {
+	if _, err := controller.NetworkByName(runconfig.DefaultDaemonNetworkMode().NetworkName()); err == nil {
+		return nil
 	}
-	return nil, nil
+
+	netOption := map[string]string{
+		winlibnetwork.NetworkName: runconfig.DefaultDaemonNetworkMode().NetworkName(),
+	}
+
+	ipamV4Conf := libnetwork.IpamConf{}
+	if config.bridgeConfig.FixedCIDR == "" {
+		ipamV4Conf.PreferredPool = defaultNetworkSpace
+	} else {
+		ipamV4Conf.PreferredPool = config.bridgeConfig.FixedCIDR
+	}
+
+	v4Conf := []*libnetwork.IpamConf{&ipamV4Conf}
+	v6Conf := []*libnetwork.IpamConf{}
+
+	_, err := controller.NewNetwork(string(runconfig.DefaultDaemonNetworkMode()), runconfig.DefaultDaemonNetworkMode().NetworkName(),
+		libnetwork.NetworkOptionGeneric(options.Generic{
+			netlabel.GenericData: netOption,
+		}),
+		libnetwork.NetworkOptionIpam("default", "", v4Conf, v6Conf, nil),
+	)
+
+	if err != nil {
+		return fmt.Errorf("Error creating default network: %v", err)
+	}
+	return nil
 }
 
 // registerLinks sets up links between containers and writes the
@@ -257,6 +401,6 @@ func restoreCustomImage(is image.Store, ls layer.Store, rs reference.Store) erro
 	return nil
 }
 
-func (daemon *Daemon) networkOptions(dconfig *Config) ([]nwconfig.Option, error) {
-	return nil, fmt.Errorf("Network controller config reload not aavailable on Windows yet")
+func driverOptions(config *Config) []nwconfig.Option {
+	return []nwconfig.Option{}
 }
